+++ lib/src/data_source/decoder/mod.rs
@@ -1,2 +1,3 @@
 pub mod build_record_batch;
+#[allow(clippy::module_name_repetitions)]
 pub mod json_decoder;