+++ lib/src/data_source/mod.rs
@@ -1,15 +1,16 @@
 #[cfg(feature = "plugin")]
 pub mod connector_plugin;
 pub mod csv;
+mod decoder;
 #[cfg(feature = "flight")]
 pub mod flight_stream;
 pub mod infer_schema;
 pub mod json;
-pub mod location_uri;
+pub mod local_fs;
+pub mod location;
 pub mod nd_json;
+pub mod object_store;
 pub mod parquet;
-mod reader;
 pub mod schema;
 mod transport;
 mod with_jsonpath;
-pub mod writer;